requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.3
Django==3.0.8
django-appconf==1.0.5
mypy-extensions==0.4.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
pathspec==0.9.0
pilkit==2.0
Pillow==8.4.0
platformdirs==2.4.0
pytz==2021.3
regex==2021.10.8
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
